Fix remove busy file

+++ start.py
@@ -42,7 +42,8 @@ def progress(thread_id):
     if os.path.exists(filename):
         with open(filename, 'rb+') as f:
             data = str(struct.unpack('f', f.read()))
-        os.remove(filename)
+        if data == 1:
+            os.remove(filename)
     else:
         data = 0
     return str(data)
